Encapsulate javacOptions inside DefaultJavaLibrary
Summary:

We want to abstract away javac specific things from
DefaultJavaLibrary, so first, we prevent anyone reaching in to its
javac options, so that, later, we can substitute a different compile
phase in. Moved a couple of tests from KnownBuildRuleTypesTest to
JavaBuckConfigTest; removed two android test variants that would have
become duplicates there.

Test Plan: CI
